Added GroupRepository, corrected Group entity, minor refactor of packages
