+++ src/main/java/com/kodilla/ecommercee/controller/GroupController.java
@@ -1,13 +1,13 @@
-package com.kodilla.ecommercee.group.controller;
+package com.kodilla.ecommercee.controller;
 
-import com.kodilla.ecommercee.group.domain.GroupDto;
+import com.kodilla.ecommercee.domain.dto.GroupDto;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
 import java.util.List;
 
 @RestController
-@RequestMapping("v1/api/groups")
+@RequestMapping("/v1/group")
 public class GroupController {
 
     @GetMapping
