+++ src/main/java/com/kodilla/ecommercee/domian/Product.java
@@ -1,6 +1,7 @@
 package com.kodilla.ecommercee.domian;
 
 
+import com.kodilla.ecommercee.domain.Group;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
